rename to "Plugin Communication"
"Plugin Development: Messages and Notifications" was unwieldy.  The suggested "Plugin Development" implied the page was about all aspects of developing a plugin, which this page isn't.  Since messages and notifications are used for communication with Notepad++, and primarily (though not solely) used by plugins, I hope "Plugin Communication" is a reasonable compromise between the differing points of view -- keeping the title short, and about "plugins", but without the extra scope of "Plugin Development".