searching: clarify named classes
per notepad-plus-plus/npp-usermanual/issues/83, need to clarify that named classes must be contained in a class

The "which must be part of a character class" from the previous version was intended to indicate the rules, but it was apparently non-obvious to some users, so this update makes it much more explicit, with examples.  Also explicitly shows the complements of those character classes.

While modifying the searching docs, also clarified that `(?-s)` does _not_ affect character classes, because that's a common misunderstanding.