Update preferences.md
per https://community.notepad-plus-plus.org/post/67045,  it wasn't clear to users (or to me) that the User Extension list for a language is theme-specific.  Added a parenthetical in the paragraph on the default and user extensions to make this more clear.

While I'm here, also clarify user-defined keywords in the stylers.xml section.