Show Non-Printable Characters
New feature in active development per https://github.com/notepad-plus-plus/notepad-plus-plus/pull/13020

Started the documentation now, even though the feature hasn't been committed, because Don wanted to know what the URL would be for the table.

The table was populated based on @ozone10's list that was previously shown in dialog box, but with the abbreviations updated based on https://en.wikipedia.org/wiki/List_of_Unicode_characters#General_Punctuation

The feature is still in-progress, so the documentation is, too.