Update Notepad replacement instructions for Windows 11
Uninstalling UWP Notepad version does not make the replacement to be
respected. AFAIK this is because there is `UseFilter` set, with which
only the exact full paths mentioned in registry subkeys are used for
replacement. So either UseFilter must be set to 0, or (IMO better) just
follow the approach taken by UWP Notepad and add Debugger to each
subkey.

Still then UWP Notepad would be run from Start Menu and when
double-clicking on text file. As UWP apps are activated in special way,
not just by running a binary, this registry trick does not work for
them. So to fix ability to run from Start Menu UWP Notepad must just be
uninstalled and then sadly you have no Notepad shortcut in Start Menu.
After uninstall you will observe that there is no application associated
with text files, not even built-in Notepad. This is because there is
another registry setting that disables using built-in Notepad to open
files, so it also must be removed.