Editing > Typing Mode (#721)
Per Forum discussion, needed to describe Insert vs Override mode on the Editing page.
closes #720

But describing that the Typing Mode is listed on the Status Bar made it clear that the manual needs a section in User Interface to describe the Status Bar.  

The Typing Mode and Status Bar sections then made it obvious that the mention of `INS/DEL` in the **Preferences > General > Status Bar > ☐ Hide** was wrong, and the better Status Bar description now eliminates the need for the brief (incorrect) description in the Preferences -- so remove the bad description and link to the new section instead.

And finally, the Status Bar description needed to mention **Search > Go to...**, which was never described in the **Searching** page.